refacto cleaning profile function
