+++ package.json
@@ -14,7 +14,7 @@
     "start": "next start -p $PORT",
     "test": "jest --testPathIgnorePatterns=e2e --runInBand",
     "test:e2e": "jest --testPathPattern=e2e --runInBand",
-    "test:e2e:local": "yarn test:e2e:delete-profile-user && yarn test:e2e:remove-screenshots && ts-node bin/e2e/prepare-e2e-tests.ts && ts-node bin/e2e/setting-sudo-ttl.ts && jest --testPathPattern=e2e --runInBand",
+    "test:e2e:local": "yarn e2e:delete-profile-user && yarn test:e2e:remove-screenshots && ts-node bin/e2e/prepare-e2e-tests.ts && ts-node bin/e2e/setting-sudo-ttl.ts && jest --testPathPattern=e2e --runInBand",
     "test:e2e:delete-profile-user": "ts-node bin/e2e/delete-profile-user.ts",
     "test:e2e:remove-screenshots": "find tests/e2e/screenshots/ -type f -name '*.png' -print -exec rm -rf '{}' \\;",
     "storybook": "start-storybook -p 6006",
@@ -28,7 +28,7 @@
     "e2e:remove-heroku-redirect-uri": "node bin/e2e/remove-heroku-redirect-uri-from-connect.mjs",
     "e2e:create-test-user": "ts-node bin/e2e/create-test-user.ts",
     "e2e:remove-test-user": "ts-node bin/e2e/remove-test-user.ts",
-    "e2e:cleaning-profile-staging": "ts-node bin/e2e/cleaning-profile-staging.ts",
+    "e2e:delete-profile-user": "ts-node bin/e2e/delete-profile-user.ts",
     "e2e:setting-sudo-ttl": "ts-node bin/e2e/setting-sudo-ttl.ts",
     "generate-connect-profile-client": "docker run --rm -v ${PWD}/:/local openapitools/openapi-generator-cli generate --input-spec /local/connect-profile-openapi.yml --generator-name typescript-axios --additional-properties supportsES6=true --output /local/connect-profile-client",
     "heroku:promote-staging-to-prod": "ts-node bin/run-promote.ts"
